+++ NEWS.md
@@ -1,6 +1,8 @@
-# crew 0.0.4.9000
+# crew 0.0.5
 
 * Use `as.character()` to make sure the `list` argument supplied to `rm()` is safe.
+* Remove the `seconds_poll_high` and `seconds_poll_low` fields and arguments.
+* Hard-code the `async_dial` field in the router.
 
 # crew 0.0.4